@@ -1,13 +1,9 @@
 import React from 'react'
-import { Link, History } from 'react-router';
+import { Link } from 'react-router';
 import SearchBar from './SearchBar';
 
 var Header = React.createClass({
-
-	mixins: [History],
-
-	render: function() {
-		
+	render() {
 		return (
 			<header className='flex-row'>
 				<Link className='nav-button fa icon-setmine click center' to='/' onlyActiveOnIndex={true} />
@@ -27,12 +23,12 @@ var Header = React.createClass({
 					<div className='center'>Blog</div>
 				</Link>*/}
 				<div className='buffer-4x'/>
-				<div className='search-bar flex-row flex-3x'>
-					<SearchBar push={this.props.push} appState={this.props.appState}/>
+				<div className='flex-row flex-3x' style={{marginRight: '1vw'}}>
+					<SearchBar/>
 				</div>
 			</header>
 		);
 	}
 });
 
-module.exports = Header;
+export default Header;
