@@ -1,15 +1,12 @@
 var express = require('express');
 var path = require('path');
-var httpProxy = require('http-proxy');
-// var jsdom = require('jsdom');
 var fs = require('fs');
-
-var proxy = httpProxy.createProxyServer();
 var app = express();
 
 var isProduction = process.env.NODE_ENV === 'production';
-var port = isProduction ? 3000 : process.env.PORT;
-var publicPath = path.resolve(__dirname, 'public');
+var port = isProduction ? 80 : 3000;
+
+app.use(express.static(__dirname + '/public'));
 
 app.use(function( req, res, next ) {
     for(var prop in req.query) {
@@ -18,25 +15,6 @@ app.use(function( req, res, next ) {
     }
     next();
 });
-
-app.use(express.static(publicPath));
-
-if (!isProduction) {
-    var bundle = require('./server/bundle.js');
-    bundle();
-
-    app.all('/build/*', function (req, res) {
-        proxy.web(req, res, {
-            target: 'http://localhost:8080'
-        });
-    });
-
-}
-
-proxy.on('error', function(e) {
-    console.log('Could not connect to proxy, please try again...');
-});
-
 
 app.get('*', function( req, res, next ) {
 
