@@ -1,22 +1,13 @@
 import React from 'react';
-import Immutable from 'immutable';
-import Router from 'react-router';
-import DocMeta from 'react-doc-meta';
-import R from 'ramda';
+import {render} from 'react-dom';
+import { Router, IndexRoute, Link, Route, History, Redirect } from 'react-router';
 
-import { IndexRoute, Link, Route, History, Redirect } from 'react-router';
-import GlobalEventHandler from './services/globalEventHandler';
-import {startFacebookSDK} from './services/loginService';
-import detectMobileService from './services/detectMobileService';
-import constants from './constants/constants';
-
+import App from './components/App';
 import Footer from './components/Footer';
-import Header from './components/Header';
-import Player from './components/Player';
 import LandingView from './components/LandingView';
 import EventsView from './components/EventsView';
-import HomeView from './components/HomeView';
-import SetsView from './components/SetsView';
+import Home from './components/Home';
+import Sets from './components/Sets';
 import SearchResultsView from './components/SearchResultsView';
 
 import UpcomingEvents from './components/UpcomingEvents';
@@ -47,192 +38,13 @@ import DMCA from './components/DMCA';
 import Setmusic from './components/Setmusic';
 import SetstoryLandingPage from './components/SetstoryLandingPage';
 import About from './components/About';
-import BlogWrapper from './components/BlogWrapper';
-
-var initialAppState = Immutable.Map({
-	currentSet: {
-		set_length: '00:00',
-		starttime: '00:00',
-		id: null
-	},
-	tracklist: [],
-	currentTrack: null,
-	sound: null,
-	playerHidden: true,
-	playing: false, //change to true once set starts playing
-	timeElapsed: 0, //update while playing
-
-	artistBrowseData: [],
-	festivalBrowseData: [],
-	mixBrowseData: [],
-	activityBrowseData: [],
-	recentBrowseData: [],
-	popularBrowseData: [],
-	landingEvents: [],
-
-	isUserLoggedIn: false,
-	user: {
-		id: 108
-	},
-	newSets: [],
-	newEvents: [],
-
-	detailId: null,
-	detailData: {
-		"sets": [],
-		"upcomingEvents": [],
-		"links": {
-			"facebook": null,
-			"twitter": null,
-			"instagram": null,
-			"soundcloud": null,
-			"youtube": null
-		}
-	},
-
-	location: {
-		label: 'DEFAULT LOCATION',
-		location: {
-			lat: 29.652175,
-			lng: -82.325856
-		}
-	},
-
-	searchResults: {
-		artists: [],
-		sets: [],
-		upcomingEvents: [],
-		tracks: []
-	}
-});
-
-var evtHandler = GlobalEventHandler(initialAppState);
-var evtTypes = evtHandler.types;
-
-var push = evtHandler.push;
-
-var App = React.createClass({
-	displayName: 'App container',
-	mixins: [History],
-
-	getInitialState() {
-		return {
-			// Let's assume that other ephemeral state
-			// MAY have to exist here.
-			appState: initialAppState
-		};
-	},
-
-	componentWillMount() {
-		this._attachStreams();
-		detectMobileService.detectMobileBrowser();
-		this.playSet();
-	},
-
-	componentDidMount() {
-		var metadataPath = window.location.pathname;
-		startFacebookSDK(push);
-	},
-
-	_attachStreams() {
-		var self = this;
-		evtHandler.floodGate.subscribe(newState => {
-			self.setState({ appState: newState });
-		});
-	},
-
-	playSet() {
-		var id = this.props.params.set;
-		var self = this;
-
-		if(!!id) {
-			this.getSetById(id)
-				.done(function(res) {
-					var set = R.head(res.payload.set);
-					var currentSet = {
-						artist: set.artist,
-						event: set.event,
-						id: set.id,
-						set_length: set.set_length,
-						songURL: set.songURL,
-						artistimageURL: set.artistimageURL,
-						starttime: '00:00'
-					};
-
-					self.getTracklist(id)
-					.done(function(res) {
-						var tracklist = res.payload.tracks;
-						push({
-							type: 'SHALLOW_MERGE',
-							data: {
-								currentSet: currentSet,
-								tracklist: tracklist,
-								currentTrack: R.head(res.payload.tracklist),
-								playing: true
-							}
-						});
-					});
-				});
-		}
-	},
-
-	getSetById(id) {
-		return (
-			$.ajax({
-				type: 'get',
-				url: constants.API_ROOT + 'set/id',
-				data: {
-					'setId': [id]
-				}
-			})
-		);
-	},
-
-	getTracklist(id) {
-		return (
-			$.ajax({
-				url: constants.API_ROOT + 'tracklist/' + id,
-				type: 'get'
-			})
-		);
-	},
-
-	render() {
-		var appState = this.state.appState;
-		var tags = [
-			{property: "description", content: "Setmine is a music app dedicated to live events! Relive past music festivals: Ultra, Coachella + more! Find upcoming shows + buy tix + listen to DJs' sets"},
-			{property: "og:site_name", content: "Setmine"},
-			// {property: "og:url", content: "https://setmine.com/metadata/" + encodeURIComponent(metadataPath.substring(1))},
-			{property: "fb:app_id", content: "648288801959503"},
-			{property: "og:description", content: "Setmine offers live music enthusiasts a new way to experience their favorite festival music.  No more struggling to find your favorite sets--we've done it all for you.  Listen to Ultra, Coachella, TomorrowWorld, and many more! Also don't forget to listen your favorite DJ's radio shows!"},
-			{property: "og:image", content: "https://setmine.com/images/setmine-logo-facebook.png"},
-			{property: "og:title", content: "Setmine | View Lineups & Play Sets | Relive Your Favorite Events"},
-			{property: "og:type", content: "website"},
-			{name: "google-site-verification", content: "T4hZD9xTwig_RvyoXaV9XQDYw5ksKEQywRkqaW-CGY4"}
-		];
-		
-		return (
-			<div id='App' className='flex-column'>
-				<DocMeta tags={tags} />
-
-				<Header appState={appState} push={push}/>
-				{
-					React.cloneElement(this.props.children, {
-						appState: appState,
-						push: push
-					})
-				}
-				<Player appState={appState} push={push} />
-			</div>
-		);
-	}
-});
+// import BlogWrapper from './components/BlogWrapper';
 
 var routes = (
 	<Route path='/' component={App}>
 		<IndexRoute component={LandingView}/>
 
-		<Route path='play/:set' component={SetsView}>
+		<Route path='play/:set' component={Sets}>
 			<IndexRoute component={Recent}/>
 			<Route path='mixes' component={Mixes}/>
 			<Route path='popular' component={Popular}/>
@@ -240,7 +52,7 @@ var routes = (
 			<Route path='activities' component={Activities}/>
 		</Route>
 
-		<Route path='sets' component={SetsView}>
+		<Route path='sets' component={Sets}>
 			<IndexRoute component={Recent}/>
 			<Route path='mixes' component={Mixes}/>
 			<Route path='popular' component={Popular}/>
@@ -248,7 +60,7 @@ var routes = (
 			<Route path='activities' component={Activities}/>
 		</Route>
 
-		<Route path='user' component={HomeView}>
+		<Route path='user' component={Home}>
 			<IndexRoute component={Favorites}/>
 			<Route path='sets' component={NewSets}/>
 			<Route path='events' component={NewEvents}/>
@@ -277,6 +89,7 @@ var routes = (
 		<Redirect from='/browse/:festival/festival' to='/festival/:festival' />
 		<Redirect from='/browse/:mix/mix' to='/mix/:mix' />
 		<Redirect from='/event/:eventID' to='/event/:eventID' />
+		<Redirect from='/offer/:offerId' to='/' />
 
 		<Route path='about' component={About}/>
 		<Route path='setstory' component={SetstoryLandingPage}/>
@@ -284,15 +97,11 @@ var routes = (
 	</Route>
 );
 
-var headMount = document.getElementById('head-mount-point');
 var bodyMount = document.getElementById('body-mount-point');
 
-import createBrowserHistory from 'history/lib/createBrowserHistory';
-var history = createBrowserHistory();
-// import createHashHistory from 'history/lib/createHashHistory';
-// var history = createHashHistory();
+import history from './services/history'
 
-React.render(
+render(
 	<Router history={history}>
 		{routes}
 	</Router>
