@@ -1,3 +1,4 @@
+var webpack = require('webpack');
 var HtmlWebpackPlugin = require('html-webpack-plugin');
 var path = require('path');
 
@@ -19,16 +20,22 @@ module.exports = {
 		extensions: ['', '.jsx', '.es6', '.js', '.scss'],
 		moduleDirectories: ['node_modules']
 	},
-	plugins: [new HtmlWebpackPlugin({
-		template: 'src/index-dev.html',
-		inject: 'body'
-	})],
-	devtool: 'source-map',
+	plugins: [
+		new HtmlWebpackPlugin({
+			template: 'src/index-dev.html',
+			inject: 'body'
+		}),
+		new webpack.ProvidePlugin({
+			'Promise': 'es6-promise',
+			'fetch': 'imports?this=>global!exports?global.fetch!whatwg-fetch'
+		}),
+	],
+	devtool: 'cheap-source-map',
 	module: {
 		loaders: [
 			{
-				test: /.jsx?$/,
-				loader: 'babel-loader!jsx-loader',
+				test: /\.jsx?$/,
+				loader: 'babel-loader',
 				exclude: /node_modules/,
 				query: {
 					presets: ['es2015', 'react']
