@@ -1,4 +1,4 @@
-var Webpack = require('webpack');
+var webpack = require('webpack');
 var HtmlWebpackPlugin = require('html-webpack-plugin');
 var path = require('path');
 
@@ -20,10 +20,16 @@ module.exports = {
 		extensions: ['', '.jsx', '.es6', '.js', '.scss'],
 		moduleDirectories: ['node_modules']
 	},
-	plugins: [new HtmlWebpackPlugin({
-		template: 'src/index.html',
-		inject: 'body'
-	})],
+	plugins: [
+		new HtmlWebpackPlugin({
+			template: 'src/index.html',
+			inject: 'body'
+		}),
+		new webpack.ProvidePlugin({
+			'Promise': 'es6-promise',
+			'fetch': 'imports?this=>global!exports?global.fetch!whatwg-fetch'
+		})
+	],
 	devtool: 'source-map',
 	module: {
 		loaders: [
